abstract out rewriteJavascript
