+++ src/flowCoffeescriptKernel/flowCoffeescriptKernel.js
@@ -1,10 +1,9 @@
 import { safetyWrapCoffeescript } from './safetyWrapCoffeescript';
 import { compileCoffeescript } from './compileCoffeescript';
 import { parseJavascript } from './parseJavascript';
-import { traverseJavascriptScoped } from './traverseJavascriptScoped';
 import { createRootScope } from './createRootScope';
 import { removeHoistedDeclarations } from './removeHoistedDeclarations';
-import { createGlobalScope } from './createGlobalScope';
+import { rewriteJavascript } from './rewriteJavascript';
 
 export function flowCoffeescriptKernel() {
   const lodash = window._;
@@ -12,48 +11,6 @@ export function flowCoffeescriptKernel() {
   const escodegen = window.escodegen;
   const esprima = window.esprima;
   const CoffeeScript = window.CoffeeScript;
-  const rewriteJavascript = sandbox => (rootScope, program, go) => {
-    let error;
-    const globalScope = createGlobalScope(rootScope, sandbox.routines);
-    try {
-      traverseJavascriptScoped([globalScope], globalScope, null, null, program, (globalScope, parent, key, node) => {
-        let identifier;
-        if (node.type === 'Identifier') {
-          // ignore var declarations
-          if (parent.type === 'VariableDeclarator' && key === 'id') {
-            return;
-          }
-          // ignore members
-          if (key === 'property') {
-            return;
-          }
-          identifier = globalScope[node.name];
-          if (!identifier) {
-            return;
-          }
-
-          // qualify identifier with '_h2o_context_'
-          parent[key] = {
-            type: 'MemberExpression',
-            computed: false,
-            object: {
-              type: 'Identifier',
-              name: identifier.object,
-            },
-            property: {
-              type: 'Identifier',
-              name: identifier.name,
-            },
-          };
-          return parent[key];
-        }
-      });
-      return go(null, program);
-    } catch (_error) {
-      error = _error;
-      return go(new Flow.Error('Error rewriting javascript', error));
-    }
-  };
   const generateJavascript = (program, go) => {
     let error;
     try {
